refactor(llm): wire anthropic adapter to the codec
Rewire the Anthropic adapter into a thin transport shell over the
`anthropic_messages` codec from the previous commit. All wire translation
now lives in the codec; the adapter owns only auth, base URL, the
streaming byte loop, and route selection.

- Resolve the three inline `provider_name == "anthropic"` request-time
  branches (auth scheme, version/beta headers, count-tokens gate,
  forced-streaming) into a single `RouteConfig` computed once per call.
  Direct Anthropic uses x-api-key + version/beta headers + count-tokens;
  every other name (Kimi-over-anthropic) is bearer auth, no anthropic
  headers, no count route, blocking-via-streaming.
- Resolve file-backed attachments via the shared `attachments::resolve`
  (introduced earlier in the series) before the sync codec `encode`,
  constructing the Anthropic policy (images + documents, not audio) inline;
  this consumes the infra, so its ahead-of-use `dead_code` allow is dropped.
- Migrate the ~70 anthropic unit tests into the codec submodules they now
  cover: encode/decode/stream/wire helper tests in place; the former
  `build_api_request` tests assert on the encoded body/headers via a
  direct-route `encode` helper; adapter-identity and count-tokens HTTP
  tests stay in the adapter shell.

Behavior-preserving: the 126 wire snapshots and full fabro-llm suite (515
tests) stay green; clippy `-D warnings` and fmt clean.
